requirements.txt
revChatGPT==*
fastapi==0.86.0
uvicorn==0.19.0
grpcio==1.50.0
grpcio-tools==1.50.0
